+++ saleor/account/utils.py
@@ -129,19 +129,6 @@ def change_user_default_address(
         set_user_default_shipping_address(user, address)
 
 
-def remove_staff_member(staff):
-    """Remove staff member account only if it has no orders placed.
-
-    Otherwise, switches is_staff status to False.
-    """
-    if staff.orders.exists():
-        staff.is_staff = False
-        staff.user_permissions.clear()
-        staff.save()
-    else:
-        staff.delete()
-
-
 def retrieve_user_by_email(email):
     """Retrieve user by email.