+++ saleor/graphql/discount/tests/benchmark/test_promotion_create.py
@@ -197,7 +197,7 @@ def test_promotion_create_order_promotion(
 
     # when
     staff_api_client.ensure_access_token()
-    with django_assert_num_queries(36):
+    with django_assert_num_queries(38):
         response = staff_api_client.post_graphql(PROMOTION_CREATE_MUTATION, variables)
 
     # then